% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{address_strings}
\alias{address_strings}
\title{address_strings}
\usage{
address_strings(frame = FALSE)
}
\arguments{
\item{frame}{Optionally return data frame. If so, will not work with clean_addresses.}
}
\value{
address_strings returns a character vector.
}
\description{
This function simply loads a character vector of common
address names and abbreviations.
}
\note{
Optionally return data frame.
}
