% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_names.R
\name{clean_schools}
\alias{clean_schools}
\title{clean_schools}
\usage{
clean_schools(.data, var, vec_of_strings = NULL)
}
\arguments{
\item{.data}{The data frame that is intended for modification.}

\item{var}{The column of school names that will be modified.}

\item{vec_of_strings}{Optional vector of character strings to replace}
}
\value{
clean_schools returns a data frame with the new column 'clean_school'
in the first position, followed by the original school name column, and finally
the rest of the columns. None are dropped. If 'var' is not character, original
data frame is returned.
}
\description{
This function removes and replaces common school pronouns
and abbreviations to prepare for matching on school name.
}
\details{
The simpleschoolsR package is collection of concise data sets and functions
designed to alleviate the burden of importing, cleaning, and
matching data sets with U.S. schools as a primary key. For example,
an analyst who wishes to join data from the National Center for Education
Statistics to data from the Civil Rights DAta Collection may be able to
use NCES IDs. However, in instances where a common ID is not available,
names and addresses may allow for matching...when they are accurate and
consistent across data frames. Unfortunately, this is rarely the case, resulting
in a frustrating and costly experience for the analyst. This package seeks to
alleviate basic issues stemming from these inconsistencies.
}
\note{
returns original data frame if selected column is not character.
}
