% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_us_suffix}
\alias{dic_us_suffix}
\title{Common abbreviations for street suffixes}
\format{
A data frame with columns:
\describe{
 \item{suf.type}{Categorized street type}
 \item{suf.input}{Common abbreviation}
 \item{suf.output}{Unified output of abbreviations}}
}
\source{
postmastr package
}
\usage{
dic_us_suffix
}
\description{
Common abbreviations for street suffixes
}
\keyword{datasets}
