% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{directional_abbs}
\alias{directional_abbs}
\title{Common abbreviations for directions}
\format{
A data frame with columns:
\describe{
 \item{dir.input}{Common directional abbreviation}
 \item{dir.output}{Unified output of abbreviations}
  }
}
\source{
postmastr package
}
\usage{
directional_abbs
}
\description{
Common abbreviations for directions
}
\keyword{datasets}
