% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intersections}
\alias{intersections}
\title{Data to help solve common issues with elements of addresses}
\format{
A data frame with columns:
\describe{
 \item{intersect.input}{An abbreviation for "at"}
 \item{intersect.output}{The word "at"}
 }
}
\source{
postmastr package
}
\usage{
intersections
}
\description{
Special character abbreviations
}
\keyword{datasets}
