% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/school_names.R
\docType{data}
\name{school_strings}
\alias{school_strings}
\title{Common school name abbreviations}
\format{
A data frame with columns:
\describe{
 \item{input}{Common abbreviation}
 \item{output}{Unified output of abbreviations}
  }
}
\usage{
school_strings

school_strings(frame = FALSE)
}
\arguments{
\item{frame}{Optionally return data frame. If so, will not work with clean_schools.}
}
\value{
school_strings returns a character vector.
}
\description{
This function simply loads a character vector of common
school names and abbreviations.
}
\note{
Optionally return data framereturns original data frame if selected column is not character.
}
\keyword{datasets}
